Move Reverb integration tests and add Reverb Github workflow
