+++ .github/workflows/reverb.yml
@@ -1,0 +1,69 @@
+name: reverb
+
+on:
+  push:
+  pull_request:
+
+env:
+  RUN_BLOCKING_TESTS: true
+
+jobs:
+  reverb:
+    runs-on: ubuntu-latest
+    timeout-minutes: 10
+
+    services:
+      redis:
+        image: redis:8
+        ports:
+          - 6379:6379
+        options: >-
+          --health-cmd "redis-cli ping"
+          --health-interval 10s
+          --health-timeout 5s
+          --health-retries 5
+
+    container:
+      image: phpswoole/swoole:6.1.4-php8.4
+
+    strategy:
+      fail-fast: true
+
+    name: Reverb Integration Tests
+
+    steps:
+      - name: Checkout code
+        uses: actions/checkout@v6
+
+      - name: Log trigger context
+        uses: ./.github/actions/log-trigger-context
+
+      - name: Setup PHP extensions
+        uses: ./.github/actions/setup-php
+
+      - name: Cache Composer dependencies
+        uses: actions/cache@v5
+        with:
+          path: /root/.composer/cache
+          key: composer-8.4-${{ hashFiles('composer.lock') }}
+          restore-keys: composer-8.4-
+
+      - name: Install dependencies
+        run: COMPOSER_MEMORY_LIMIT=-1 composer install --prefer-dist -n -o
+
+      - name: Start Reverb test servers
+        env:
+          REDIS_HOST: redis
+          REDIS_PORT: 6379
+        run: |
+          REVERB_SERVER_PORT=19510 php tests/Integration/Reverb/server.php &
+          sleep 2
+          REVERB_SERVER_PORT=19511 REVERB_SCALING_ENABLED=true php tests/Integration/Reverb/server.php &
+          sleep 3
+
+      - name: Execute Reverb integration tests
+        env:
+          REDIS_HOST: redis
+          REDIS_PORT: 6379
+          ENGINE_TEST_SERVER_HOST: 127.0.0.1
+        run: vendor/bin/phpunit tests/Integration/Reverb
