+++ tests/Integration/Reverb/ReverbRedisIntegrationTestCase.php
@@ -2,13 +2,13 @@
 
 declare(strict_types=1);
 
-namespace Hypervel\Tests\Reverb\Integration;
+namespace Hypervel\Tests\Integration\Reverb;
 
 /**
  * Base test case for Reverb integration tests with Redis scaling enabled.
  *
  * Requires a running Redis-enabled Reverb test server on port 19511.
- * Start it with: REVERB_SERVER_PORT=19511 REVERB_SCALING_ENABLED=true php tests/Reverb/Integration/server.php
+ * Start it with: REVERB_SERVER_PORT=19511 REVERB_SCALING_ENABLED=true php tests/Integration/Reverb/server.php
  *
  * Tests auto-skip when the server is unavailable (InteractsWithServer
  * from the parent class handles this). Redis availability is implicitly
